# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tech Mentor ERP & POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /publisher-series:
    get:
      operationId: listPublisherSeries
      tags: [book-catalog]
      parameters:
        - { name: brandId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: List of publisher series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublisherSeries"
    post:
      operationId: createPublisherSeries
      tags: [book-catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherSeriesInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherSeries"
  /publisher-series/{id}:
    patch:
      operationId: updatePublisherSeries
      tags: [book-catalog]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherSeriesInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublisherSeries"
    delete:
      operationId: deletePublisherSeries
      tags: [book-catalog]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /book-classes:
    get:
      operationId: listBookClasses
      tags: [book-catalog]
      responses:
        "200":
          description: List of book classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookClass"
    post:
      operationId: createBookClass
      tags: [book-catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookClass"
  /book-classes/{id}:
    delete:
      operationId: deleteBookClass
      tags: [book-catalog]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /book-subjects:
    get:
      operationId: listBookSubjects
      tags: [book-catalog]
      responses:
        "200":
          description: List of book subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookSubject"
    post:
      operationId: createBookSubject
      tags: [book-catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookSubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookSubject"
  /book-subjects/{id}:
    delete:
      operationId: deleteBookSubject
      tags: [book-catalog]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # CATEGORIES
...
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [brands]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    delete:
      operationId: deleteBrand
      tags: [brands]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # INVENTORY
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    patch:
      operationId: updateBranch
      tags: [branches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # POS SALES
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, role, name, status, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        role: { type: string }
        status: { type: string }
        branchId: { type: ["integer", "null"] }
        createdAt: { type: string }
...
    Brand:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        productCount: { type: integer }
        createdAt: { type: string }
    BrandInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    BrandUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }

    # INVENTORY
...
    Branch:
      type: object
      required: [id, name, address, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        isMain: { type: boolean }
        status: { type: string }
        createdAt: { type: string }
    BranchInput:
      type: object
      required: [name, address]
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        isMain: { type: boolean }
    BranchUpdate:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string }

    # SALE (POS)
...
    PublisherSeries:
      type: object
      required: [id, brandId, name, createdAt]
      properties:
        id: { type: integer }
        brandId: { type: integer }
        brandName: { type: ["string", "null"] }
        name: { type: string }
        createdAt: { type: string }
    PublisherSeriesInput:
      type: object
      required: [brandId, name]
      properties:
        brandId: { type: integer }
        name: { type: string }
    BookClass:
      type: object
      required: [id, name, sortOrder, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string }
    BookClassInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        sortOrder: { type: integer }
    BookSubject:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: string }
    BookSubjectInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
...